+++ src/keywords.js
@@ -1,4 +1,5 @@
 module.exports = {
   add: "add",
   git: "git",
+  update: "-u",
 };